+++ .gitignore
@@ -37,3 +37,6 @@ cmake-build-debug
 
 # MacOS
 .DS_STORE
+
+# Logger files
+*.log
