+++ include/utils.hpp
@@ -88,34 +88,4 @@ private:
 template<class T>
 T* Singleton<T>::instance_ = nullptr;
 
-class OutFileStream : public NonCopyable {
-public:
-    OutFileStream() = default;
-    OutFileStream(const fs::path& path) { this->set_path(path); }
-
-    void set_path(const fs::path& path) {
-        if (!fs::exists(path) || fs::is_regular_file(path))
-            path_ = path;
-    }
-    const fs::path& path() const { return path_; }
-
-    void open(std::ios_base::openmode mode = std::ios::out) { output_.open(path_, mode); }
-    bool is_open() const { return output_.is_open(); }
-    void close() { if (this->is_open()) output_.close(); }
-
-    OutFileStream& operator<<(std::ostream& (*os)(std::ostream&)) {
-        output_ << os;
-        return (*this);
-    }
-    template <typename T>
-    OutFileStream& operator<<(const T& data) {
-        output_ << data;
-        return (*this);
-    }
-
-private:
-    fs::path path_;
-    std::ofstream output_;
-};
-
-}
+} // utils
