+++ src/main.cpp
@@ -7,5 +7,5 @@ int main() {
     LOG(logger::info) << "Hello, world!";
     LOGI << "Short form";
 
-    return 0;
+  return 0;
 }
